core_django/requirements.txt
# Django Core
django
djangorestframework
psycopg2-binary
django-cors-headers
requests
pyjwt
